extract_pdf_to_csv/requirements.txt
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.3
et_xmlfile==2.0.0
msoffcrypto-tool==5.4.2
numpy==2.2.6
olefile==0.47
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.2.1
pycparser==2.22
pypdfium2==4.30.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2
